plotting added

+++ init_database.sql
@@ -1,7 +1,7 @@
 -- sensor_readings
 CREATE TABLE IF NOT EXISTS sensor_readings (
     id integer PRIMARY KEY,
-    ldr float,
-    sunpanel float,
+    ldr integer,
+    sunpanel integer,
     text_datetime text NOT NULL
 );
